requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
firecrawl-py==2.7.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.61
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
MarkupSafe==3.0.2
multidict==6.4.4
narwhals==1.40.0
nest-asyncio==1.6.0
numpy==2.2.6
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.45.1
tavily-python==0.7.2
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
